+++ app/src/main/java/org/simple/clinic/home/patients/PatientsTabEffect.kt
@@ -4,10 +4,6 @@ import org.simple.clinic.patient.businessid.Identifier
 
 sealed class PatientsTabEffect
 
-object OpenEnterOtpScreen : PatientsTabEffect()
-
-data class OpenPatientSearchScreen(val additionalIdentifier: Identifier?) : PatientsTabEffect()
-
 object RefreshUserDetails : PatientsTabEffect()
 
 object LoadUser : PatientsTabEffect()
@@ -16,18 +12,24 @@ object LoadInfoForShowingApprovalStatus : PatientsTabEffect()
 
 data class SetDismissedApprovalStatus(val dismissedStatus: Boolean) : PatientsTabEffect()
 
-object ShowUserWasApproved : PatientsTabEffect()
-
-object HideUserAccountStatus : PatientsTabEffect()
-
-object OpenScanBpPassportScreen : PatientsTabEffect()
-
 object LoadNumberOfPatientsRegistered : PatientsTabEffect()
-
-object OpenTrainingVideo : PatientsTabEffect()
 
 object LoadInfoForShowingAppUpdateMessage : PatientsTabEffect()
 
 object TouchAppUpdateShownAtTime : PatientsTabEffect()
 
-object ShowAppUpdateAvailable : PatientsTabEffect()
+sealed class PatientsTabViewEffect : PatientsTabEffect()
+
+object OpenEnterOtpScreen : PatientsTabViewEffect()
+
+data class OpenPatientSearchScreen(val additionalIdentifier: Identifier?) : PatientsTabViewEffect()
+
+object ShowUserWasApproved : PatientsTabViewEffect()
+
+object HideUserAccountStatus : PatientsTabViewEffect()
+
+object OpenScanBpPassportScreen : PatientsTabViewEffect()
+
+object OpenTrainingVideo : PatientsTabViewEffect()
+
+object ShowAppUpdateAvailable : PatientsTabViewEffect()