+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Bump logback to v1.2.5
 - Show appointment results in contact patient sheet for patient with no phone number
 - Change `PatientsTabScreen` to use Mobius view effects
+- Bump lint to v30.0.1
 
 ### Changes
 - [In Progress: 19 Aug 2021] UI improvements for medical history screen